+++ src/utils.rs
@@ -27,6 +27,7 @@ pub fn get_path(path: &str) -> Result<PathBuf, Box<dyn Error>> {
 
 pub fn print_table(builder: Builder) {
     let table = builder
+        .index()
         .build()
         .with(Style::modern_rounded())
         .with(Alignment::center())