+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # BUAA Cli: Powerful tool to Make BUAA Great Again
 
+## [v0.3.4] - 2025-12-31
+
+- Fix
+  - `tes` command now work. But there are some **WARNINGS**
+    > **Warning!**: Due to the poor design of the evaluation system server,
+    > using this may cause the evaluation button on the web page to become unclickable.
+    > But don't worry, the evaluation data has been submitted correctly.
+    > If you want to view the evaluation results on the web page,
+    > you can remove the 'disabled' attribute of the button in the browser console,
+    > and you'll be able to click it.
+    > Or you might wait a little longer, and it may return to normal.
+- Chore
+  - Update `buaa_api`
+  - Cargo fmt
+
 ## [v0.3.3] - 2025-12-30
 
 - Fix